MultiSwingView can display multiple views depending on data contents. Various tweaks.
Added direct record view to allow a ledger to just display the elements
in a list as one field.
Got rid of names in single target reaction
Made RefGeneral and ValGeneral covariant, and added OptionalReaction
for ListVars and Vars that only applies if a function returns
Some(value), does nothing for None
Adapted BoxesDemo and BoxSpec for these changes.
Made GraphZoomerAxis have Double values, not Int (no idea why it was
Int!)
Removed commented code from popup button (was left over from old popup
implementation, not needed with dialogs)
Added optional minimum width to SheetBuilder
Added MultiSwingView, this can create new SwingViews to deal with
polymorphic data. It will create either a new view per class of data,
or per instance of data. Both systems are written to avoid memory leaks
(both data items and their views will be GCed in the per-instance case,
in the per-class case data items will be GCed, but one view will be
retained forever per class seen at runtime).

+++ boxes/src/boxes/graph/Graph.scala
@@ -905,7 +905,7 @@ class GraphBox(fill:RefGeneral[Color, _], outline:RefGeneral[Color, _], enabled:
 }
 
 case class GraphZoomerAxis(
-    requiredRange:Ref[Option[(Int, Int)]] = Var(None),
+    requiredRange:Ref[Option[(Double, Double)]] = Var(None),
     padding:Ref[Double] = Var(0.05),
     minSize:Ref[Double] = Var(0.01)
 )
@@ -921,8 +921,8 @@ class GraphZoomer(
       case None => {
         //We have no data bounds, so use the axes required ranges,
         //or 0 to 1 in each axis if there are none.
-        val xRange = xAxis().requiredRange().getOrElse((0, 1))
-        val yRange = yAxis().requiredRange().getOrElse((0, 1))
+        val xRange = xAxis().requiredRange().getOrElse((0d, 1d))
+        val yRange = yAxis().requiredRange().getOrElse((0d, 1d))
         Area(Vec2(xRange._1, yRange._1), Vec2(xRange._2, yRange._2)).normalise
       }
       case Some(area) => {
